+++ utoipa/CHANGELOG.md
@@ -13,6 +13,7 @@ to look into changes introduced to **`utoipa-gen`**.
 
 ### Changed
 
+* Finalize actix-web utoipa bindings (https://github.com/juhaku/utoipa/pull/1160)
 *  Update utoipa-gen version
 
 ## 5.1.0 - Oct 16 2024